+++ R/within_network.R
@@ -52,51 +52,75 @@ is_within_any <- function(address, network) {
 
 #' Network membership of other IP networks
 #'
-#' These functions check if an IP network is a subnet or supernet
-#' of another network.
+#' @description
+#' `overlaps()` checks for any overlap between two networks.
 #'
-#' @param network1 An \code{\link{ip_network}} vector
-#' @param network2 An \code{\link{ip_network}} vector
+#' `is_subnet()` and `is_supernet()` check if one network is a true
+#' subnet or supernet of another network.
+#'
+#' @param network An \code{\link{ip_network}} vector
+#' @param other An \code{\link{ip_network}} vector
 #' @return A logical vector
 #'
 #' @examples
-#' net1 <- ip_network("192.168.1.0/24")
-#' net2 <- ip_network("192.168.1.128/30")
+#' net1 <- ip_network("192.168.1.128/30")
+#' net2 <- ip_network("192.168.1.0/24")
 #'
-#' is_subnet(net2, net1)
+#' overlaps(net1, net2)
+#'
+#' is_subnet(net1, net2)
 #'
 #' is_supernet(net1, net2)
+#' @seealso
+#' Use [is_within()] to check if an \code{\link{ip_address}} is within
+#' an \code{\link{ip_network}}.
 #' @name network_in_network
 NULL
 
 #' @rdname network_in_network
 #' @export
-is_subnet <- function(network1, network2) {
+overlaps <- function(network, other) {
   assertthat::assert_that(
-    is_ip_network(network1),
-    is_ip_network(network2)
+    is_ip_network(network),
+    is_ip_network(other)
   )
 
   # vector recycling
-  args <- vec_recycle_common(network1, network2)
-  network1 <- args[[1L]]
-  network2 <- args[[2L]]
+  args <- vec_recycle_common(network, other)
+  network <- args[[1L]]
+  other <- args[[2L]]
 
-  is_within_wrapper(network_address(network1), network2) & (prefix_length(network1) >= prefix_length(network2))
+  is_within_wrapper(network_address(network), other) | is_within_wrapper(network_address(other), network)
 }
 
 #' @rdname network_in_network
 #' @export
-is_supernet <- function(network1, network2) {
+is_subnet <- function(network, other) {
   assertthat::assert_that(
-    is_ip_network(network1),
-    is_ip_network(network2)
+    is_ip_network(network),
+    is_ip_network(other)
   )
 
   # vector recycling
-  args <- vec_recycle_common(network1, network2)
-  network1 <- args[[1L]]
-  network2 <- args[[2L]]
+  args <- vec_recycle_common(network, other)
+  network <- args[[1L]]
+  other <- args[[2L]]
 
-  is_within_wrapper(network_address(network2), network1) & (prefix_length(network2) >= prefix_length(network1))
+  is_within_wrapper(network_address(network), other) & (prefix_length(network) >= prefix_length(other))
+}
+
+#' @rdname network_in_network
+#' @export
+is_supernet <- function(network, other) {
+  assertthat::assert_that(
+    is_ip_network(network),
+    is_ip_network(other)
+  )
+
+  # vector recycling
+  args <- vec_recycle_common(network, other)
+  network <- args[[1L]]
+  other <- args[[2L]]
+
+  is_within_wrapper(network_address(other), network) & (prefix_length(other) >= prefix_length(network))
 }